# shared/contracts/openapi/batch-downloader-api.yaml
openapi: 3.1.0
info:
  title: Batch Downloader API
  version: 1.0.0
  description: API pública versionada para identidad, catálogo y trabajos de descarga.
  license:
    name: GPL-3.0-only
    identifier: GPL-3.0-only
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /users/me/downloads:
    get:
      operationId: listOwnDownloadHistory
      summary: Listar el historial registrado al alcanzar READY
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Historial paginado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadHistoryPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /download-jobs:
    post:
      operationId: createDownloadJob
      summary: Crear un trabajo de descarga
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDownloadJobRequest'
      responses:
        '202':
          description: Trabajo persistido y encolado.
          headers:
            Set-Cookie:
              description: Se emite para el primer trabajo anónimo y solo contiene un token aleatorio del navegador.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Ninguna aplicación solicitada tiene una fuente verificable o el bundle excede el límite.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Cuota superada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /download-jobs/{jobId}:
    get:
      operationId: getDownloadJob
      summary: Consultar un trabajo de descarga
      security: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Estado global e items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadJob'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: cancelDownloadJob
      summary: Cancelar un trabajo de descarga
      security: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '202':
          description: Cancelación aceptada de forma idempotente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadJob'
        '404':
          $ref: '#/components/responses/NotFound'
  /download-jobs/{jobId}/events:
    get:
      operationId: streamDownloadJobEvents
      summary: Seguir el progreso de un trabajo
      security: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Flujo SSE hasta alcanzar un estado terminal.
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /download-jobs/{jobId}/file:
    get:
      operationId: getDownloadJobFile
      summary: Obtener el ZIP de un trabajo terminado
      security: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '303':
          description: Redirección a una URL firmada válida durante cinco minutos.
          headers:
            Location:
              required: true
              schema:
                type: string
                format: uri
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: BATCH_SESSION
    internalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Entrada no válida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Sesión ausente o inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: La sesión o el estado de la cuenta no permiten la operación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: El estado actual impide la operación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    CreateDownloadJobRequest:
      type: object
      additionalProperties: false
      properties:
        appIds:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
        bundleId:
          type: string
          minLength: 1
        operatingSystems:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [windows, linux, macos]
        notifyWhenReady:
          type: boolean
          default: false
      oneOf:
        - required: [appIds]
        - required: [bundleId]
    DownloadJobStatus:
      type: string
      enum: [QUEUED, RESOLVING, DOWNLOADING, PACKAGING, READY, PARTIAL, MANUAL_ONLY, FAILED, CANCELLED, EXPIRED]
    DownloadItem:
      type: object
      required: [id, appId, appName, status, bytesDownloaded]
      properties:
        id:
          type: string
          format: uuid
        appId:
          type: string
          format: uuid
        appName:
          type: string
          minLength: 1
        officialPageUrl:
          type: [string, 'null']
          format: uri
        status:
          type: string
          enum: [QUEUED, RESOLVING, DOWNLOADING, COMPLETED, FAILED, CANCELLED]
        bytesDownloaded:
          type: integer
          format: int64
        sha256:
          type: [string, 'null']
        errorCode:
          type: [string, 'null']
    DownloadJob:
      type: object
      required: [id, status, progress, requestedCount, acceptedCount, omittedCount, failureCode, items, createdAt, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/DownloadJobStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        requestedCount:
          type: integer
          minimum: 1
        acceptedCount:
          type: integer
          minimum: 1
        omittedCount:
          type: integer
          minimum: 0
        failureCode:
          type: [string, 'null']
        items:
          type: array
          items:
            $ref: '#/components/schemas/DownloadItem'
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...
    DownloadHistoryItem:
      type: object
      additionalProperties: false
      required: [appId, appName, jobId, downloadedAt]
      properties:
        appId:
          type: string
          format: uuid
        appName:
          type: string
        slug:
          type: [string, 'null']
        iconUrl:
          type: [string, 'null']
          format: uri
        jobId:
          type: string
          format: uuid
        downloadedAt:
          type: string
          format: date-time
    DownloadHistoryPage:
      type: object
      additionalProperties: false
      required: [data, page, pageSize, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DownloadHistoryItem'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true